--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -9,7 +9,6 @@ include("framework.jl")
 eye_(n) = Matrix{Int64}(I, n, n)
 
 my_tests = [
-            "test_delayed_systems",
             "test_statespace",
             "test_transferfunction",
             "test_zpk",
@@ -28,6 +27,7 @@ my_tests = [
             "test_synthesis",
             "test_partitioned_statespace",
             "test_demo_systems",
+            "test_delayed_systems",
             ]