--- a/test/test_timeresp.jl
+++ b/test/test_timeresp.jl
@@ -37,6 +37,39 @@ yd, td, xd = lsim(sysdfb, zeros(501), t, x0=x0)
 @test y ≈ yd
 @test x ≈ xd
 
+# Error for nonuniformly spaced vector
+@test_throws ErrorException lsim(sys, [1, 2, 3, 4], [0, 1, 1, 2])
+
+# Error for mismatch of sample rates of discrete-system and signal
+@test_throws ErrorException lsim(sysd, [1, 2, 3, 4], 0:0.2:0.6)
+
+# Test for problem with broadcast dimensions
+@test lsim(sys, zeros(5), 0:0.2:0.8)[1][:] == zeros(5)
+
+
+# lsim for discrete-time complex-coefficient systems
+
+# Complex system, real input signal
+G1 = ss(1.0im, 1)
+@test lsim(G1, [1, 2], 0:1)[1][:] == 1.0im*[1, 2]
+@test impulse(G1, 4)[1][:] == [1.0im; zeros(4)]
+@test step(G1, 4)[1][:] == fill(1.0im, 5)
+
+G2 = ss(1.0im, 1, 1, 0, 1)
+@test lsim(G2, ones(3), 0:2)[1][:] == [0.0, 1, 1 + im]
+@test impulse(G2, 4)[1][:] == [0.0, 1, im, -1, -im]
+@test step(G2, 4)[1][:] == [0.0, 1, 1+im, 1+im-1, 0]
+
+# Real system, complex input signal
+@test lsim(ss(1, 1), [1.0, 1.0im], 0:1)[1][:] == [1.0, 1.0im]
+@test lsim(ss(1.0, 1), [1.0, 1.0im], 0:1)[1][:] == [1.0, 1.0im]
+@test lsim(ss(1, 1, 1, 0, 1), [1.0, 1im, 1], 0:2)[1][:] == [0.0, 1, 1 + im]
+
+# Complex system, complex input signal
+@test lsim(ss(1.0im, 1), [1im, 2], 0:1)[1][:] == [-1, 2im]
+@test lsim(ss(1.0im, 1, 1, 0, 1), [1.0, 1im, 1], 0:2)[1][:] == [0.0, 1, 2im]
+
+
 
 #Test step and impulse Continuous
 t0 = 0:0.05:2
@@ -119,6 +152,10 @@ y, t2, x = step(G, 2)
 
 #Make sure t was never changed
 @test t0 == t
+
+
+
+
 
 
 @testset "Simulators" begin