`ltitr` was re-written to use in-place operations and allow different numeric types of the input data.

This allows complex-valued simulations (closes #264) and makes things faster.

Minor changes were done to `lsim`:
* Avoid checking `_issmooth` when simulating discrete-time systems (`_issmooth` fails for complex signals) 
* `x0` was restricted to be a `Vector` to avoid broadcast problems, this seems quite reasonable and should be good for performance? Or should it be `AbstractVector`?
* Also added a check that the time vector is regularly spaced, if it isn't then the result will be horribly wrong.

It will be possible to improve things by quite a bit when we can dispatch on continuous or discrete. For example, `ltitr` could probably be replaced with the discrete-time version of `lsim`.