@baggepinnen 2020-04-07
- `AbstractVector` is always to be preferred, you never know when there's a `FillArray` etc. appearing as an initial state.
- Do you have some before and after benchmarks to share? (It's very satisfying to see successful perf. improvements :D )