@olof3 2020-04-07
Some benchmarks of `ltitr`, quite many allocations from the slices, but looked through this with @mfalt , and didn't see any obvious improvements. 5-argument `mul!` gives another 10 percent or so.

```
Benchmark 1:
old:   37.197 μs (802 allocations: 89.20 KiB)
new:   16.504 μs (1199 allocations: 54.75 KiB)
Benchmark 2:
old:   49.565 μs (802 allocations: 134.52 KiB)
new:   27.535 μs (1199 allocations: 65.73 KiB)
```

```julia
using BenchmarkTools

function ltitr(A::AbstractMatrix{T}, B::AbstractMatrix{T}, u::AbstractVecOrMat,
        x0::VecOrMat=zeros(T, size(A, 1))) where T
    N = size(u, 1)
    x = similar(A, size(A, 1), N)
    for i=1:N
        x[:,i] = x0
        x0 = A * x0 + B * u[i,:]
    end
    return transpose(x)
end


@views function ltitr_new(A::AbstractMatrix, B::AbstractMatrix, u::AbstractVecOrMat,
        x0::AbstractVecOrMat=zeros(eltype(A), size(A, 1)))

    T = promote_type(LinearAlgebra.promote_op(LinearAlgebra.matprod, eltype(A), eltype(x0)),
                      LinearAlgebra.promote_op(LinearAlgebra.matprod, eltype(B), eltype(u)))

    n = size(u, 1)

    # Transposing u allows column-wise operations, which apparently is faster.
    # See issue for discssuing the orientation of the input u and the output x.
    ut = transpose(u)

    x = Matrix{T}(undef, size(A, 1), n)
    tmp = Vector{T}(undef, size(A, 1)) # temp vector for storing B*u[:,k]

    x[:,1] .= x0
    for k=1:n-1
        mul!(x[:,k+1], A, x[:,k])

        mul!(tmp, B, ut[:,k]) # these two lines could be replaced with 5-argument mul! (>= v1.3)
        x[:, k+1] .+= tmp
    end
    return transpose(x)
end

nx = 3; nu = 1
A = rand(nx,nx)
B = rand(nx,nu)
u = rand(200,nu)
x0 = rand(nx)

println("Benchmark 1:")
print("old: ")
@btime ltitr(A, B, u, x0)
print("new: ")
@btime ltitr_new(A, B, u, x0)

nx = 10; nu = 5
A = rand(nx,nx)
B = rand(nx,nu)
u = rand(200,nu)
x0 = rand(nx)

println("Benchmark 2:")
print("old: ")
@btime ltitr(A, B, u, x0)
print("new: ")
@btime ltitr_new(A, B, u, x0)
```