@baggepinnen 2020-04-07
Nice, that's quite an okay speedup. The extra allocations are expected and could potentially be removed using UnsafeArrays.jl, even though I very seldom see that working in practice. The small allocations are amortized for larger ste vectors so this is still very nice 😃