@olof3 2020-04-11
Seems like it is possible to shave off another 30-40 % or so if all the B*u operations are done as a matrix--matrix multiplication (avoiding many of the view allocations?), although the code becomes less readable.

Continuation from the benchmark above.

```julia
# with 5-argument mul!
@views function ltitr_newest1(A::AbstractMatrix, B::AbstractVecOrMat, u::AbstractVecOrMat,
        x0::AbstractVecOrMat=zeros(eltype(A), size(A, 1)))

    T = promote_type(LinearAlgebra.promote_op(LinearAlgebra.matprod, eltype(A), eltype(x0)),
                      LinearAlgebra.promote_op(LinearAlgebra.matprod, eltype(B), eltype(u)))

    n = size(u, 1)

    x = Matrix{T}(undef, size(A, 1), n)
    x[:,1] .= x0

    mul!(x[:, 2:end], B, transpose(u[1:end-1, :])) # Initialize the result matrix x with B*u

    for k=1:n-1
        mul!(x[:,k+1], A, x[:,k], 1, 1) # muladd Ax[,k] to B*u that is already stored x[:,k+1]
    end
    return transpose(x)
end


# without 5 argument mul!
@views function ltitr_newest2(A::AbstractMatrix, B::AbstractVecOrMat, u::AbstractVecOrMat,
        x0::AbstractVecOrMat=zeros(eltype(A), size(A, 1)))

    T = promote_type(LinearAlgebra.promote_op(LinearAlgebra.matprod, eltype(A), eltype(x0)),
                      LinearAlgebra.promote_op(LinearAlgebra.matprod, eltype(B), eltype(u)))

    n = size(u, 1)

    tmp = Vector{T}(undef, size(A, 1)) # temp vector for storing B*u[:,k]

    x = Matrix{T}(undef, size(A, 1), n)

    mul!(x[:, 2:end], B, transpose(u[1:end-1, :])) # Initialize x with B*u

    x[:,1] .= x0
    for k=1:n-1
        mul!(tmp, A, x[:,k])
        x[:,k+1] .+= tmp
    end
    return transpose(x)
end
```


```
Benchmark 1 (nx=3, nu=1, nt=200):
old:       37.937 μs (802 allocations: 89.20 KiB)
new:      17.248 μs (1199 allocations: 54.75 KiB)
newest1:  9.905 μs (403 allocations: 23.66 KiB)
newest2:  12.718 μs (603 allocations: 33.09 KiB)

Benchmark 2 (nx=10, nu=5, nt=200):
old:      49.805 μs (802 allocations: 134.52 KiB)
new:      27.397 μs (1199 allocations: 65.73 KiB)
newest1:  15.670 μs (403 allocations: 34.59 KiB)
newest2:  19.249 μs (603 allocations: 44.08 KiB)

Benchmark 3 (nx=50, nu=5, nt=200):
old:      117.297 μs (803 allocations: 393.84 KiB)
new:      80.540 μs (1200 allocations: 128.52 KiB)
newest1:  65.013 μs (404 allocations: 97.05 KiB)
newest2:  74.126 μs (604 allocations: 106.86 KiB)
```