@olof3 2020-10-08
Yes, I think this could be pulled. 

I saw that you @baggepinnen had made some changes to the promotion in `lsim` in another PR, I removed that change as I believe that that is already handled in this PR, making it redundant. Tests seem to pass, so hopefully I didn't break anything.

However, we should revisit `lsim` sometime soon. With the introduction of `TimeEvolution` the `lsim` function could be made a lot cleaner by moving a lot of the functionality to the `lsim(::AbstractStateSpace{Discrete}, ...)` function (that you coincidentally just introduced in an other PR). Then the continuous-time version could call the discrete version when appropriate. That would avoid some logic and clutter. 

But perhaps that could be for another time.