@mfalt 2020-10-08
@JuliaControlBot test-plots

(Updating the reference plots, since the bot was broken)